pyproject.toml
[tool.poetry]
name = "keepintouch"
version = "0.1.0"
description = "Webhook-based telegram bot, that helps stay in touch with hundreds of people."
authors = ["Vladimir Vyazovetskov <erhosen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
python-telegram-bot = "^13.13"
humanize = "^4.2.3"
python-dateutil = "^2.8.2"
psycopg2-binary = "^2.9.3"
django-dotenv = "^1.4.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
requests-mock = "^1.9.3"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
